refactor: extract mounted artifact capture
